Reduce API surface by making fields private
Make allmost all fields and types private to simplify API, and now that
it can only be built via parsing, then String() and Binary() methods
will not return error.

If the structures have inconsistent data, that clearly indicates a bug
that should be fixed, hence functions will panic with a clear error
message; this should simplify bug reporting and fixing.

Signed-off-by: Gustavo Serrano <gustavo.serrano@cloudsoda.io>